Crate substudy

Tools for studying foreign languages using subtitles. All APIs are currently experimental or unstable, but if you'd like me to stabilize things, please get in touch.

For further details about how to use substudy, see the main GitHub page.

Modules

align
Align two subtitle files.

clean
Tools for cleaning up subtitle files and getting them into some kind of normalized format.

contexts
Iteration in context, with the previous and next value, if available.

decode
Decode text in a wide variety of character encodings.

err
Error-handling for this library.

export
Interfaces to various spaced repetition systems.

lang
Naming and identifying languages. We use

merge
Tools for merge multiple subtitles into one.

srt
SRT-format subtitle support.

time
Tools for working with time.

video
Tools for working with video files.