Struct substudy::contexts::Context

pub struct Context<T> {
    pub prev: Option<T>,
    pub curr: T,
    pub next: Option<T>,
}

An item in a slice, possibly with the item before or after it.

Fields

prev

The previous item, if any.

curr

The current item.

next

The next item, if any.

Methods

impl<'a, T> Context<&'a T>

fn map<U, F>(&self, f: F) -> Context<U> where F: FnMut(&'a T) -> U

Transform a Context<&'a T> to Context<U> by applying f to all embedded values.

impl<'a, T: Clone> Context<&'a T>

fn cloned(&self) -> Context<T>

Convert a Context<&T> to Context<T>.

impl<'a, T> Context<&'a Option<T>>

fn flatten(&self) -> OptContext<&'a T>

If we have a Context<&Option<T>>, that would make the type of prev a truly obnoxious Option<&Option<T>>. This function will flatten this into Option<&T>, using the OptContext helper type.