Module substudy::export

Interfaces to various spaced repetition systems.

Structs

Exporter
Information about media file and associated subtitles that the user wants to export.

LanguageResources
Information about a specific language.

Functions

export_csv
Export the video and subtitles as a CSV file with accompanying media files, for import into Anki.

export_review
Export the video and subtitles as a web page in "reviewable" format.

export_tracks
Export the video as a set of tracks.

os_str_to_string
Take a platform-specific pathname fragment and turn it into a regular Unicode string.