Struct substudy::lang::Lang

pub struct Lang {
    // some fields omitted
}

A language identifier.

Methods

impl Lang

fn iso639(code: &str) -> Result<Lang>

Specify a language using an ISO 639-1, -2/T or -2/B code. We know that the same language is sometimes represented by more than one code, and we do our best to treat equivalent codes as the same language.

use substudy::lang::Lang;
assert_eq!(Lang::iso639("en").unwrap(), Lang::iso639("eng").unwrap());
assert!(Lang::iso639("en").unwrap() != Lang::iso639("fr").unwrap());
assert!(Lang::iso639("abcd").is_err());

fn as_str(&self) -> &str

Get the normalized language code as a &str. Prefers ISO 639-1 codes when possible, and -2/T if that's the best it can do.

use substudy::lang::Lang;
assert_eq!("en", Lang::iso639("en").unwrap().as_str());
assert_eq!("en", Lang::iso639("eng").unwrap().as_str());

fn for_text(text: &str) -> Option<Lang>

Try to determine the language of text. We return None unless we're pretty sure.