Struct substudy::srt::Subtitle

pub struct Subtitle {
    pub index: usize,
    pub period: Period,
    pub lines: Vec<String>,
}

A single SRT-format subtitle, minus some of the optional fields used in various versions of the file format.

Fields

index

The index of this subtitle. We should normalize these to start with 1 on output.

period

The time period during which this subtitle is shown.

lines

The lines of text in this subtitle.

Methods

impl Subtitle

fn to_string(&self) -> String

Return a string representation of this subtitle.

fn plain_text(&self) -> String

Return a plain-text version of this subtitle.