Module substudy::video

Tools for working with video files.

Structs

Extraction
Information about what kind of data we want to extract.

Fraction
A wrapper around Ratio with custom serialization support.

Id3Metadata
Information about an MP3 track (optional).

Stream
An individual content stream within a video.

Video
Represents a video file on disk.

Enums

CodecType
Individual streams inside a video are labelled with a codec type.

ExtractionSpec
What kind of data do we want to extract, and from what position in the video clip?