Struct substudy::video::Video

pub struct Video {
    // some fields omitted
}

Represents a video file on disk.

Methods

impl Video

fn new(path: &Path) -> Result<Video>

Create a new video file, given a path.

fn file_name(&self) -> &OsStr

Get just the file name of this video file.

fn file_stem(&self) -> &OsStr

Get just the file stem of this video file, stripped of any extensions.

fn streams(&self) -> &[Stream]

List all the tracks in a video file.

fn audio_for(&self, lang: Lang) -> Option<usize>

Choose the best audio for the specified language.

fn extract(&self, extractions: &[Extraction]) -> Result<()>

Perform a list of extractions as efficiently as possible. We use a batch interface to avoid making too many passes through the file. We assume that the extractions are sorted in temporal order.