Crate substudy

Tools for studying foreign languages using subtitles. All APIs are currently experimental or unstable, but if you'd like me to stabilize things, please get in touch.

For further details about how to use substudy, see the main GitHub page.

Modules

align
Align two subtitle files.

clean
Tools for cleaning up subtitle files and getting them into some kind of normalized format.

decode
Decode text in a wide variety of character encodings.

err
Error-handling for this library.

merge
Tools for merge multiple subtitles into one.

srt
SRT-format subtitle support.